+++ apps/elf-api/tests/http/sharing.rs
@@ -6,21 +6,21 @@ use serde_json::Value;
 use tower::util::ServiceExt as _;
 use uuid::Uuid;
 
-use crate::{TEST_AGENT_A, TEST_AGENT_B, TEST_PROJECT_ID, TEST_TENANT_ID};
+use crate::helpers::{self, TEST_AGENT_A, TEST_AGENT_B, TEST_PROJECT_ID, TEST_TENANT_ID};
 use elf_api::{routes, state::AppState};
 
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn sharing_visibility_requires_explicit_project_grant() {
-	let Some((test_db, qdrant_url, collection)) = crate::test_env().await else {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
 		return;
 	};
-	let config = crate::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
 	let state = AppState::new(config).await.expect("Failed to initialize app state.");
 	let app = routes::router(state.clone());
 	let note_id = Uuid::new_v4();
 
-	crate::insert_note(
+	helpers::insert_note(
 		&state,
 		note_id,
 		"project_shared",
@@ -82,14 +82,14 @@ async fn sharing_visibility_requires_explicit_project_grant() {
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn core_blocks_are_explicitly_attached_and_separate_from_archival_search() {
-	let Some((test_db, qdrant_url, collection)) = crate::test_env().await else {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
 		return;
 	};
-	let config = crate::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
 	let state = AppState::new(config).await.expect("Failed to initialize app state.");
 	let app = routes::router(state.clone());
 	let admin_app = routes::admin_router(state.clone());
-	let private_block_id = crate::create_core_block(
+	let private_block_id = helpers::create_core_block(
 		&admin_app,
 		"agent_private",
 		"private_operating_context",
@@ -98,7 +98,7 @@ async fn core_blocks_are_explicitly_attached_and_separate_from_archival_search()
 	.await;
 	let note_id = Uuid::new_v4();
 
-	crate::insert_note(
+	helpers::insert_note(
 		&state,
 		note_id,
 		"agent_private",
@@ -108,10 +108,11 @@ async fn core_blocks_are_explicitly_attached_and_separate_from_archival_search()
 	.await;
 
 	let (status, _) =
-		crate::attach_core_block(&admin_app, private_block_id, TEST_AGENT_A, "private_only").await;
-	let before_sessions = crate::search_session_count(&state).await;
-	let blocks = crate::get_core_blocks(&app, TEST_AGENT_A, "private_only").await;
-	let after_sessions = crate::search_session_count(&state).await;
+		helpers::attach_core_block(&admin_app, private_block_id, TEST_AGENT_A, "private_only")
+			.await;
+	let before_sessions = helpers::search_session_count(&state).await;
+	let blocks = helpers::get_core_blocks(&app, TEST_AGENT_A, "private_only").await;
+	let after_sessions = helpers::search_session_count(&state).await;
 
 	assert_eq!(status, StatusCode::OK);
 	assert_eq!(before_sessions, after_sessions);
@@ -125,30 +126,38 @@ async fn core_blocks_are_explicitly_attached_and_separate_from_archival_search()
 	assert!(blocks["items"][0]["audit_history"].as_array().expect("audit history").len() >= 2);
 	assert!(!blocks.to_string().contains("archival note must not appear"));
 
-	let b_private = crate::get_core_blocks(&app, TEST_AGENT_B, "private_only").await;
+	let b_private = helpers::get_core_blocks(&app, TEST_AGENT_B, "private_only").await;
 
 	assert_eq!(b_private["items"].as_array().expect("items array").len(), 0);
 
-	let shared_block_id = crate::create_core_block(
+	let shared_block_id = helpers::create_core_block(
 		&admin_app,
 		"project_shared",
 		"shared_operating_context",
 		"Constraint: Shared core context requires explicit project grant and attachment.",
 	)
 	.await;
-	let (denied_status, _) =
-		crate::attach_core_block(&admin_app, shared_block_id, TEST_AGENT_B, "private_plus_project")
-			.await;
+	let (denied_status, _) = helpers::attach_core_block(
+		&admin_app,
+		shared_block_id,
+		TEST_AGENT_B,
+		"private_plus_project",
+	)
+	.await;
 
 	assert_eq!(denied_status, StatusCode::FORBIDDEN);
 
-	crate::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
+	helpers::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
 
-	let (shared_status, _) =
-		crate::attach_core_block(&admin_app, shared_block_id, TEST_AGENT_B, "private_plus_project")
-			.await;
-	let b_shared = crate::get_core_blocks(&app, TEST_AGENT_B, "private_plus_project").await;
-	let b_wrong_profile = crate::get_core_blocks(&app, TEST_AGENT_B, "private_only").await;
+	let (shared_status, _) = helpers::attach_core_block(
+		&admin_app,
+		shared_block_id,
+		TEST_AGENT_B,
+		"private_plus_project",
+	)
+	.await;
+	let b_shared = helpers::get_core_blocks(&app, TEST_AGENT_B, "private_plus_project").await;
+	let b_wrong_profile = helpers::get_core_blocks(&app, TEST_AGENT_B, "private_only").await;
 
 	assert_eq!(shared_status, StatusCode::OK);
 	assert_eq!(b_shared["items"].as_array().expect("items array").len(), 1);
@@ -162,18 +171,18 @@ async fn core_blocks_are_explicitly_attached_and_separate_from_archival_search()
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn org_shared_note_is_visible_across_projects() {
 	let Some((test_db, app, state, note_id)) =
-		crate::org_shared_note_is_visible_across_projects_fixture().await
+		helpers::org_shared_note_is_visible_across_projects_fixture().await
 	else {
 		return;
 	};
-	let list_before_json = crate::list_org_shared_notes_as_reader(&app).await;
+	let list_before_json = helpers::list_org_shared_notes_as_reader(&app).await;
 
 	assert_eq!(list_before_json["items"].as_array().expect("Missing items array.").len(), 0);
 
-	crate::publish_org_shared_note_as_reader_can_see(&app, note_id).await;
+	helpers::publish_org_shared_note_as_reader_can_see(&app, note_id).await;
 
 	let grant_upsert_payload = serde_json::json!({ "grantee_kind": "project" }).to_string();
-	let grant_upsert_response = crate::post_with_authorization_and_json_body(
+	let grant_upsert_response = helpers::post_with_authorization_and_json_body(
 		&app,
 		"/v2/spaces/org_shared/grants",
 		"Bearer admin-token",
@@ -185,7 +194,7 @@ async fn org_shared_note_is_visible_across_projects() {
 
 	assert_eq!(grant_upsert_response.status(), StatusCode::OK);
 
-	crate::assert_note_visible_to_project_reader(&app, &state, note_id).await;
+	helpers::assert_note_visible_to_project_reader(&app, &state, note_id).await;
 
 	test_db.cleanup().await.expect("Failed to cleanup test database.");
 }
@@ -193,15 +202,15 @@ async fn org_shared_note_is_visible_across_projects() {
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn sharing_project_grant_enables_agent_access_to_shared_note() {
-	let Some((test_db, qdrant_url, collection)) = crate::test_env().await else {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
 		return;
 	};
-	let config = crate::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
 	let state = AppState::new(config).await.expect("Failed to initialize app state.");
 	let app = routes::router(state.clone());
 	let note_id = Uuid::new_v4();
 
-	crate::insert_note(
+	helpers::insert_note(
 		&state,
 		note_id,
 		"project_shared",
@@ -209,7 +218,7 @@ async fn sharing_project_grant_enables_agent_access_to_shared_note() {
 		"Fact: shared note with explicit grant.",
 	)
 	.await;
-	crate::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
+	helpers::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
 
 	let response = app
 		.clone()
@@ -267,15 +276,15 @@ async fn sharing_project_grant_enables_agent_access_to_shared_note() {
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn sharing_publish_creates_scope_and_grant_visibility() {
-	let Some((test_db, qdrant_url, collection)) = crate::test_env().await else {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
 		return;
 	};
-	let config = crate::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
 	let state = AppState::new(config).await.expect("Failed to initialize app state.");
 	let app = routes::router(state.clone());
 	let note_id = Uuid::new_v4();
 
-	crate::insert_note(
+	helpers::insert_note(
 		&state,
 		note_id,
 		"agent_private",
@@ -284,7 +293,7 @@ async fn sharing_publish_creates_scope_and_grant_visibility() {
 	)
 	.await;
 
-	let initial_grant_count = crate::active_project_grant_count(&state, TEST_AGENT_A).await;
+	let initial_grant_count = helpers::active_project_grant_count(&state, TEST_AGENT_A).await;
 
 	assert_eq!(initial_grant_count, 0);
 
@@ -316,7 +325,7 @@ async fn sharing_publish_creates_scope_and_grant_visibility() {
 	assert_eq!(publish_json["note_id"], note_id.to_string());
 	assert_eq!(publish_json["space"], "team_shared");
 
-	let after_grant_count = crate::active_project_grant_count(&state, TEST_AGENT_A).await;
+	let after_grant_count = helpers::active_project_grant_count(&state, TEST_AGENT_A).await;
 
 	assert_eq!(after_grant_count, 1);
 
@@ -377,15 +386,15 @@ async fn sharing_publish_creates_scope_and_grant_visibility() {
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_GRPC_URL (or ELF_QDRANT_URL) to run."]
 async fn sharing_revoke_project_grant_removes_visibility() {
-	let Some((test_db, qdrant_url, collection)) = crate::test_env().await else {
+	let Some((test_db, qdrant_url, collection)) = helpers::test_env().await else {
 		return;
 	};
-	let config = crate::test_config(test_db.dsn().to_string(), qdrant_url, collection);
+	let config = helpers::test_config(test_db.dsn().to_string(), qdrant_url, collection);
 	let state = AppState::new(config).await.expect("Failed to initialize app state.");
 	let app = routes::router(state.clone());
 	let note_id = Uuid::new_v4();
 
-	crate::insert_note(
+	helpers::insert_note(
 		&state,
 		note_id,
 		"project_shared",
@@ -393,9 +402,9 @@ async fn sharing_revoke_project_grant_removes_visibility() {
 		"Fact: shared note for revoke test.",
 	)
 	.await;
-	crate::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
+	helpers::insert_project_scope_grant(&state, TEST_AGENT_A, TEST_AGENT_A).await;
 
-	let grant_count_before = crate::active_project_grant_count(&state, TEST_AGENT_A).await;
+	let grant_count_before = helpers::active_project_grant_count(&state, TEST_AGENT_A).await;
 
 	assert_eq!(grant_count_before, 1);
 
@@ -440,7 +449,7 @@ async fn sharing_revoke_project_grant_removes_visibility() {
 
 	assert_eq!(revoke_response.status(), StatusCode::OK);
 
-	let grant_count_after = crate::active_project_grant_count(&state, TEST_AGENT_A).await;
+	let grant_count_after = helpers::active_project_grant_count(&state, TEST_AGENT_A).await;
 
 	assert_eq!(grant_count_after, 0);
